requirements.txt
flask
clickhouse-connect
numpy
matplotlib
kafka-python